tools/pdf-to-word/pyproject.toml
[project]
name = "pdf-to-word"
version = "0.1.0"
description = "Convert scanned PDFs to Word documents for text-to-speech"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "pymupdf",
    "python-docx",
]
